Remove layout attribute from queue items
An item's layout is mainly required for freezing the job graph and in
case a following item needs it as a fallback layout.

To prevent components other than the pipeline manager from accessing the
layout, we will remove the layout as an attribute from the queue item
and store it in an internal cache of the manager.

Queue items will reference their (dynamic) layout via the layout's UUID.

This also prepares us for the scale-out scheduler where we have to
re-calculate a layout on a scheduler if it is not available in the
cache, as we are not storing the layout in Zookeeper due to it's size.

Since the project metadata is needed by the executor client, which
doesn't have access to the layout anymore, the metadata is now also
available via the job graph.

Change-Id: I93c7a3932fbf9ad8915d1d3d1fff9682778b28f8