Remove layout attribute from queue items
We want to restrict layout access to job freezing so that fewer
actions (like creating a build request after a job is frozen) require
access to it.

To prevent components other than the pipeline manager from accessing the
layout, we will remove the layout as an attribute from the queue item
and store it in an internal cache of the manager.

Queue items will reference their (dynamic) layout via the layout's UUID.
An item's layout UUID will change if any of it's input to the layout
creation changed. Those inputs are the tenant layout and list of items
ahead. This means that during a re-enqueue and in case of a gate reset
we will set the layout UUID of an item back to None.

This also prepares us for the scale-out scheduler where we have to
re-calculate a layout on a scheduler if it is not available in the
cache, as we are not storing the layout in Zookeeper due to it's size.

Since the project metadata is needed by the executor client, which
should not access to layout anymore, the metadata is now available via
the job graph.

Change-Id: I93c7a3932fbf9ad8915d1d3d1fff9682778b28f8