Web UI: make more filters selectable in build, buildset searches
When searching builds or buildsets, make some fields selectable instead
of free-form:

* Project
* Pipeline
* Job (builds only)

If the amount of options is not over 1000, enable typeahead search for
these fields (any more would incur performance issues in the browser).

Alter the behavior of the big (X) button in selects to be more intuitive:

* if the user typed some text in the input field, clear the input field.
* if the input field is empty, clear all filters set for the select.

Add pipelines to the redux store.

Change-Id: I710a91fa2013c2bde0293c4f36d0e582de47b38a