Update patch set 11
Patch Set 11:

PS11 is an updated version of PS8. I started the work on that after pushing PS8, but I didn't think that anybody else would touch this change in the meantime 😄 It already addresses most of the comments and I will take care of the remaining ones in another patch set. The same applies to the rebase - which is still necessary for this patchset.

Patch-set: 11