client: fix REST autohold-list response
The current REST client code takes the autohold API response and
modifies it into a format that the display function can't work with.
I couldn't exactly pinpoint how it got this way, but it seems
emperically that the RPC response is the same as the API response;
i.e. we can remove this munging and the display function can handle
both.

 zuul --zuul-url=https://zuul.opendev.org/ autohold-list --tenant openstack

is an example of this response.

Change-Id: I1087041f1054244130f1b5a68a0282742d6581d7