Use component registry in fingergw routing
This uses the component registry rather than gearman to perform
fingergw routing lookups.  It also adjusts the logic for routing
to match the latest version of the spec, where unzoned fingergw
process are expected to route to zoned fingergws if they exist
(because the unzoned fingergw might be a public gateway outside
of the zone).

Change-Id: I2f9fed03159db59cc4e496802b9dab05f746e1a2