Close fingergw connection handler on input termination
When a connection to the fingergw is terminated with no input, the
recv() method returns the empty string after poll returns an event
with the input flag set.

Handle that case by raising a BrokenPipeError which is already handled
by the calling code as a remote hang-up.

This has a noticeable impact on the test latency.

Change-Id: Ia56c94bc1e764b2f0f8ce3d6994689a3e2f7daff