Spec: move from OFTC to Matrix
This describes the reasoning and process for moving from IRC on OFTC
to Matrix.

If we obtain consensus on this spec, that indicates that as a community
we desire the outcome described in this spec.  The next step would then
be to create an infra-spec to obtain consensus that OpenDev is willing
to help achieve that and the technical details for doing so.

To be clear, the Zuul community doesn't get to decide what OpenDev will
do.  Instead we're going to decide what we want to ask of OpenDev, and
what we're willing to volunteer to help accomplish.

Change-Id: I3b4c0f9e3aa8e5b1464f196c124c38bd0ee0a480