Combine fingergw certificate options
This combines the client and server certificate options to make
typical deployments simpler.  The same certificate will be used by
a fingergw acting as a client or a server.

A new option is added to tell fingergw to use the cert only for
client use; that way a fingergw can act as an unencrypted end-user
gateway while still able to connect to encrypted servers.

The options are renamed to tls_* to match zookeeper; once gearman
is removed, we will have no ssl_* options.

Documentation and a release note for TLS fingergw support is added.

Change-Id: If3e445336de4644a5303f2ecc7c4a27e4320d042