Fix test race with executor stats
The stats for the number of executors online/accepting are now
calculated from the component registry.  The variable used internally
in the executor to record whether it is accepting or not has an
incorrect initial value.  Executors accept work as soon as they start,
then periodically the governor runs and re-evaluates.  At that time
the internal flag, and now component registry, are updated.

Tests that check the emitted stats can race depending on whether they
run long enough for the governor to run.  By correcting the initial
value, we will have predictable and correct stats output.

Change-Id: I6b84efb9292b6cb6fa439a1a46d47ee0ce1c4254