Update patch set 1
Patch Set 1:

E.g. external to Zuul deployment is asynchronously started by Zuul job on 2nd +2 and it is desired to abort it with another job if vote is either deleted or set to 0.
Basically logically this is equivalent to setting vote to 0 and any use case you can think of for this case is applicable to vote deletion.

Patch-set: 1