zuul-stream: drop ARA bindep install
We dropped installing ARA here when
I5ce1385245c76818777aa34230786a9dbaf723e5 converting things to using
the Zuul managed versions of Ansible (which installs ARA).  Since we
dropped the required-projects, I think the bindep here has been
silently just not doing anything.

We also stop gzipping the output -- it makes it fairly useless as
you'd have to bulk download the whole thing to unzip it.  I think this
was an artifact of the old logs.openstack.org days.

Needed-By: https://review.opendev.org/c/openstack/project-config/+/777675
Change-Id: I727a088c3176d13b483a0c28f39cbc0f51dad3f6