Gitlab: filter repository branches with protected branch rules
Gitlabs protected branch api does not return all branches that are
protected. Instead it returns a list of patterns of branches that are
protected. For example there could be a rule to make all release
branches protected with 'release/*' which would currently cause zuul to
try to load configuration from the non-existent branch 'release/*.

Fix this by requesting all branches and then filtering with the
protected branch rules.

Change-Id: I276d6c576b4f52c6463f792ef2314d6c787af1c7