Update patch set 1
Patch Set 1:

With my Zuul contributor hat on I am happy to join just about any communication platform that I can tie into weechat (my preferred client). There is a weechat plugin for matrix which means this should be easy and straightforward for me. No concerns there.

With my OpenDev hat on I'm concerned that this adds a new burden to our shrinking team. A huge chunk of my time in the last 6 months or so has been devoted to upgrading services and servers that people rely on to avoid interruption as much as possible and cleaning up any services which are unused. I worry that this plan will saddle me with another set of communication tools that I will end up being responsible for in a year or two when those initially interested are unable to continue supporting them.

For this reason I think we should strongly consider not asking OpenDev to run a homeserver. I don't know what the terms are for having Element run one but that would be my preference. Additionally, we should probably consider converting all bots to matrix if this move happens and have them talk to the OFTC bridge for other OpenDev users rather than running multiple sets of bots.

I suspect all of this is doable and that would address my primary concerns.

Patch-set: 1
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>