Update patch set 1
Patch Set 1:

Jens, I tried to address the motivation in the spec (I listed seven "Benefits Provided by Matrix", none of which were "fancy new features", and I listed six "Reasons to Move", only one of which was "fancy new features").

If that's all you got out of what I wrote, then I did a very poor job of communicating, I'm sorry.

If i had to pick one thing to counter the idea that it's about fancy new features, it's this:

Using matrix makes it easy for all users to benefit from the always-on connections that only advanced IRC users like you or I take for granted.  And this while still being able to use a variety of clients with open standards, and communicate with others over IRC networks.  This is an opportunity to lift everyone up to the same level.

Honestly, I'd be fine with stopping at "let's use OFTC but everyone should use Matrix to talk to OFTC" except that if we ever require nick registration again (which is not at all far-fetched based on our experience over the weekend) then we end up with double-authentication for matrix users, and that's arguably even more confusing than just using IRC.

And granted, it's not exactly the same.  I changed clients from irssi to weechat.  But I'm happy with the functionality that provides; I'm not missing much.  It's really rare that we can find a compromise position that works so well for everyone.

There are a lot of good reasons to stick with IRC, and I've been advocating for it for years.  But I believe that Matrix has reached a point where it outshines IRC in almost every way, and if we want users to use an open system at all, we need to start using Matrix.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>