Codify branch matchers and refs
Since time immemorial the branch matcher code has handled matching
explicit refs so that jobs can specify on what tags or other refs they
should run.  But we never documented it.  And the only test is a
trivial unit test:

  tests.unit.test_change_matcher.TestBranchMatcher.test_matches_returns_true_on_matching_ref

Add a more robust test that ensures that branch matchers can be used
to selectively run jobs based on tag names, and update the documentation
to describe this.

Change-Id: If836544550b1b87ecb389a1ab39c393b97736551