github: show query response on missing fields
The nested_data function returns None on missing keys, but most of the
parsing here assumes that prior steps have returned correct data.  For
the most part this would be correct, but per the linked story
sometimes you can get something back that doesn't parse.  Move the
parsing under an exception catch for AttributeError and raise a more
specific exception that shows the bad data for debugging.

Change-Id: Ic03b44956d31ad9c5681e7e268bcecaadb4cf1d0
Story: #2008940