Move build request cleanup from executor to scheduler
The scheduler is a better home for the build request cleanup because
the cleanup process should run across all zones, and the executor
api object in the executors are tied to specific zones, so they
have an incomplete cache.

It also means that if all the executors go offline, the schedulers
can still cancel and reset the builds, so it doesn't look like
dead builds are still running.

Since we now have two cleanup actions, instead of making a thread
for each, we use APScheduler to run the cleanup functions on an
interval.  We use ZK locks to make sure that they aren't running
on more than one scheduler.

Change-Id: I0245b71f31aae9616d8e65d27c63b25b2c27f815