github: fallback to api_token when can't find installation
graphql queries (I77be4f16cf7eb5c8035ce0312f792f4e8d4c3e10) require
authentication, so we can not really run in anonymous mode.

If the app is setup (app_id in config) but we aren't installed with
permissions on the project we're looking up, then fall back to using a
specified api_token.

The error-level message is changed, because this seems like a fairly
common scenario when you're Depends-On: pull requests from repos you
aren't installed for.  In this case we just want a note that we're
falling back to the API key.

We raise an exception if neither the installation token or api_token
gives us authorization; this is better than the graphql queries
failing with 401 errors later.

Documentation is updated, a changelog added and several test
configuration files are padded with the required info.

Story: #2008940
Change-Id: I2107aeafc55591eea790244701567569fa6e80d4