Consistently use UTC timestamps in fake_gitlab
This is necessary to ensure that comparisons for updates to merge
requests function properly. Otherwise depending on your timezone you
could end up with an updated_at timestamp that is older than the
created_at timestamp preventing actions for updates from happening.

In particular this was breaking the
test_merge_request_updated_builds_aborted test for those of us running
the unittests on systems without UTC clocks.

Change-Id: I133bedf5dabe0f18e30598c3eb394c0a12b0ee0a