Cleanup Zuul's stdout/stderr output
This is primarily an issue in the unittests, but we also cleanup a
problem with output in ansible package installation verification.

There are two types of issue we address here. The first is unittests
calling print(). We replace those with log.info() calls to keep this
information from adding noise to the stestr output. The second is
subprocess.run() not capturing output so it ends up on stdout/stderr. In
this case we update use of subprocess.run() to capture the output, then
log/error appropriately if the return code is not 0.

Change-Id: I22650bf9495d3fe71bdf4a2dec5d9b3f30116188