Refactor config/tenant (re-)loading
In a multi-scheduler deployment we don't want to have a global layout
lock in Zookeeper. Instead, we will use a read/write lock per tenant.
This allows use to continue processing pipelines for other tenants in
the meantime.

In respect of that, this change refactors the config loading so that we
have a single interface that performs config (re-)loading on a
per-tenant basis. This is also important as later on a scheduler startup
will no longer require a full reconfiguration.

With the refactoring, the configloader will always (re-)load the config
of a specific tenant as requested by the scheduler. If a full
reconfiguration should be performed the scheduler will create a new
Abide instance before loading any tenant. For the case of smart or
tenant reconfigurations we will re-use the existing Abide instance
instead of copying everything over from the old to the new instance.

Change-Id: I6b7afd34203b74d5a6524e5277055a813147ff9e