Fix layout cache and adding a project to a pipeline
When reporting a failure on a change that adds a project to a
pipeline, we must be careful to distinguish a failure that users
want to see versus a failure that should not be reported (if Zuul
can't merge a given change, it can't know if that change adds
the project or not, so it falls back to the parent layout to
decide if the project participates in the pipeline).

We recently reduced the number of cases where we generate a full
dynamic layout so that they are only created when freezing a job
graph of the current change, or if needed to do so for a following
change.  Any time there is a reconfiguration, the layout cache is
cleared and layouts are regenerated only if they need to be to meet
the above conditions.

With the following sequence of events, we can produce erroneous
behavior:

1) Enqueue a change which, via in-repo config, adds a project to
a pipeline.
2) While jobs are running for that change, reconfigure.
3) When the jobs complete and the change reports, there will be
no layout local to the queue item and Zuul will fall back to the
current pipeline layout where the project is not in the pipeline,
and will decide not to report.

The presence of jobs on a queue item is a de-facto indication that
the dynamic layout for that item included the project in the
pipeline.  That's rather efficient, so let's check that first,
then fall back on the layout cache for the remaining cases.

There is one case that isn't handled by this:

* If an item *ahead* is the item that adds the project to the pipeline.
* And this item has a merge failure.
* And there has been a reconfiguration since the item ahead was
  enqueued.

In that case, Zuul will still proceed through the original code path
where it finds no layout with the project in the pipeline at the time
of reporting.  To correct that, we would need to either recompute the
layout, or we would need to determine that the item is in the pipeline
earlier and preserve that information on the queue item so it persists
across reconfigurations.

Change-Id: I7a9a2005eb57c4e2fd398e616c31ba5035e8d965