Fix reporting merge failure behind change adding ppc
Change I7a9a2005eb57c4e2fd398e616c31ba5035e8d965 corrected an edge case
where a reconfiguration would clear a cached layout, and, on already
enqueued changes where a project was being added to a pipeline, the
absence of that layout would cause Zuul not to report since it could
not determine whether the project was part of the pipeline.

The fix for that suggested we should examine the case where a change
with a merge conflict is behind a successful change which adds a project
to a pipeline.  In that case, we should fall back to the layout of the
item ahead, but the item ahead is no longer in the queue.  We should,
however, have the item ahead's layout in the cache and referenced by
UUID, but we need to set our failing item's layout uuid in that case.

To correct this, we call getLayout for the side effect of saving the
layout uuid on the item in both of the cases where we otherwise
short-circuit layout generation on merge or config errors.

This case doesn't need a reconfiguration to trigger it -- an item
with no jobs behind an item with a dynamic layout will cause the item
ahead to regenate its layout, so it's not subject to the bug fixed in
I7a9a200.  Instead, this case requires the item ahead to succeed so
that it's not present in the queue when the item behind reports.

Change-Id: I7b4d1596262b8b1b5c300f44401aafe923e718b3