web: update to latest xterm; disabled selection overrides
Currently when I try to select any text on the streaming console it
goes a white-on-white color and you can't see the selected text.

Using the sample console on the xterm.js page it correctly uses a
light-grey selection that you can see through.  This doesn't appear to
set any theme or css settings to achieve this.

Upgrade to the latest release.  This requires a few changes to the way
things are imported and as far as I can tell, just calling the fit
addon fit() function on resize seems to make things work as expected.

This removes any overrides for selection related things; the console
now highlights in a light grey correctly.

Change-Id: I9699e6d9abe82a20791770a1a8f169d65be27bc9