web: update to latest xterm; disabled selection overrides
Currently when I try to select any text on the streaming console it
goes a white-on-white color and you can't see the selected text.

Using the sample console on the xterm.js page it correctly uses a
light-grey selection that you can see through.  This doesn't appear to
set any theme or css settings to achieve this.  So upgrade to the
latest release, and remove any overrides.

Change-Id: I9699e6d9abe82a20791770a1a8f169d65be27bc9