Fix unknown job detection
If geard loses track of a job, this code is supposed to detect it,
but we've moved classes, so we can't use dunderscores anymore,
we need to change it to a single underscore.

This also corrects the mirror situation where the scheduler is
disconnected from the gearman server; that would eventually fall
through to the lost build handler, but since we get an unambiguous
disconnect event, let's restore the handling of that.

A test case for each of these is added.

Change-Id: I41a926c488717f8bc4cd8a173beaa9458e7cb6e2