Update patch set 15
Patch Set 15:

PS15 implements the stuff we discussed in our meeting yesterday. I also made this change work now (tests are green) by keeping most of the code from the acceptNodes() method on the scheduler side. I did that by splitting up the acceptNodes() method into two. I've updated the commit message with a more detailed explanation. @Jim Please let me know what you think about this approach. The only part that might be a little confusing now are the changes in the test_nodepool.py file where you can see the split between acceptNodeRequest() and acceptNodes(). Because the latter one is now not an "atomic" method anymore that validates the NodeRequest, locks the nodes and directly deletes the request, but those steps are not split into two methods (which reflect the work split between scheduler and executor).

Patch-set: 15