Refactor pipeline processing in run handler
With per-pipeline event queues in place we can now refactor the run
handler to process a pipeline's events right before processing the
pipeline queues.

This way we will only loop once over the tenants and its pipelines. It
also reduces the places in the code where we have to lock the tenants
and pipelines later on.

Change-Id: I657498a2ecac09e34e5a054464a70d298a69cab0