Update patch set 2
Patch Set 2:

PS2 updates the test to remove zookeeper configuration but without fixing the CLI. It seems like to enable using the tenant-conf-check in a check job (without a ZooKeeper running) we could either:

- modify the Scheduler and the configloader to skip the cache for this use-case
- add a mock cache in the zuul.cmd.client

Which one would you prefer?

Note that this is a blocker preventing us from upgrading to Zuul 4.5.0, unless we disable the zuul tenant config check job.

Patch-set: 2