Move parent provider determination to pipeline manager
Moving the parent provider determination into the pipeline manager
allows us to remove the buildset and job objects from the NodeRequest
constructor. This way we can fully serialize the NodeRequest to
ZooKeeper and restore it without missing important information.

To ensure that NodePool doesn't strip those information when it fulfills
the NodeRequest, we use the new "requestor_data" field which is
implemented in [1].

The priority attribute is only used to determine the ZooKeeper path when
the node request is submitted. Thus, we can move this attribute out of
the class and provide it as additional parameter to the
submitNodeRequest call (and the related update callbacks).

To make this work, we also have to look up the buildset by its UUID from
the active tenants and pipelines when the NodesProvisioned event is
handled in the scheduler. Something similar was already done for
handling the other result events as well.

[1]: https://review.opendev.org/c/zuul/nodepool/+/798746/

Depends-On: https://review.opendev.org/c/zuul/nodepool/+/798746/

Change-Id: Id794643dcf26b0565499d20adba99d3b0518fdf1