Remove gearman job from build when finished
We don't need the gearman job object on the builds after a build has
been finished. Tracemalloc indicates that roughly half of the
allocated memory in our scheduler is occupied by serialized data in
gearman jobs. Thus clean the gearman jobs from the builds as soon as
they are finished since the build object has the same livetime as a
buildset.

Change-Id: I9bdf5817b655e6560c0cdea4581add8a526880c5