Freeze secrets on job level instead of playbook level
Currently the secrets are frozen and serialized per playbook. However
every secret is attached to every playbook on a job variant so we can
waste a considerable amount of memory on job variants that use
multiple secrets and multiple playbooks. Reduce the memory footprint
by moving the secrets freezing to the job variant level.

Change-Id: I3e2edde2d05ff6533db7fafac713be4c02684f14