Intern strings loaded from zuul config
Zuul configuration can contain a lot of duplicated strings all over
the place. This is especially true when projects have a large amount
of config and projects with many branches. Since strings are immutable
we can leverage the sys.intern [1] function to deduplicate the strings
we load when parsing the config and safe memory. As a side effect this
can improve the performance of comparisons and dictionary lookups.

[1] https://docs.python.org/3/library/sys.html#sys.intern

Change-Id: I46e4a4111239915c2d2f8e15f41c24f01b40895e