Re-register components after ZK reconnect
An entry in the component registry is an ephemeral znode (so that
we can tell if a component disconnects).  However, we never re-add
the entry if the component reconnects.  We only register components
in their init method.

For some components this could simply mean that the system info would
be incomplete.  For others, such as executors, it could mean that
the scheduler thinks there are no executors available in a zone.

To correct this, we add an onReconnect handler and call the register()
method on the component from it.

An alternative approach would be to have each component's main loop
re-register if it detects that a ZK error has occurred, but this
method seems both simpler and should be harmless even if it's slightly
out of sync with the component's main loop.

In addition to the unit test here, this will be exercised by functional
testing in the TestShedulerZones.test_executor_disconnect method
once we switch to putting build requests in ZK.

Change-Id: I96e93dd256eee546bed091a91b93b887b069e216