Add debug info for merger component test
We're seeing a race in test_merger_component and it's difficult
to tell the event sequence from the logs since there are multiple
component registries involved.  Add the registry object id to the
log line in the callback so we can identify which registry sees
which events.

Also, add log the list of components at the end of the test if it
fails so that we can see what the assertion saw.

Finally, increase the timeout for the assertion, in case this is
a simple case of contention on the test nodes.

Change-Id: Ie9595ac6764139011d12ba622a32b88659cd2a6f