Replace TreeCache in component registry
We found that Kazoo's TreeCache recipe can report a created event
after a node has been deleted.  That renders it unsuitable for
our use where we need to know if a component has dropped out of
the cluster.

This replaced the TreeCache with a pair of child/data watches in
much the same way as the Executor API.  It also adds a new unit
test which is useful for testing the component registry api in
isolation.

This change also e-registers components after ZooKeeper reconnection
events.

An entry in the component registry is an ephemeral znode (so that
we can tell if a component disconnects).  However, we never re-add
the entry if the component reconnects.  We only register components
in their init method.

For some components this could simply mean that the system info would
be incomplete.  For others, such as executors, it could mean that
the scheduler thinks there are no executors available in a zone.

To correct this, we add an onReconnect handler and call the register()
method on the component from it.

An alternative approach would be to have each component's main loop
re-register if it detects that a ZK error has occurred, but this
method seems both simpler and should be harmless even if it's slightly
out of sync with the component's main loop.

In addition to the unit test here, this will be exercised by functional
testing in the TestShedulerZones.test_executor_disconnect method
once we switch to putting build requests in ZK.

Because the ComponentRegistry creates new Component objects to represent
data that it reads from ZK, we must ensure those objects are created in
"read-only" mode, so that if the connection that backs the registry is
reset, these objects don't re-register as if they were the actual
components.

Finally, the timeout is increased in the assertions for the component
tests, and some extra debug info is added.  This is leftover from an
earlier change to add the debugging info needed to identify the issue
this patch fixes.  They seem like good changes to keep.

Change-Id: I96e93dd256eee546bed091a91b93b887b069e216