Update patch set 23
Patch Set 23:

Maybe we should compress and shard for both this and the config cache?

Also, if I were to hazard a guess, I'd guess that your params are so large due to repo state (lots of branches), so this will get a lot better when we put a BuildSet object in zk and put a single copy of the repo state on that.

Patch-set: 23