Shard BuildRequest parameters
It's possible for the build request parameters to get quite large,
so this uses the sharding API to split them across multiple znodes.

Since the executor is going to attempt to start processing a build
request as soon as it exists, and the natural place for the sharded
parameters is underneath the build request, if we created the request
first and then added the sequence znodes for the parameters, the
executor may try to start processing the build request before the
parameters are written.

To resolve this, the sharding API is updated so that it can accept
not only a zk client, but also a zk transaction (which behaves like
a client with some restrictions).  This means we can create the build
request and all of the sharded parameter nodes in one atomic
transaction.

One of the restrictions is that we can't pass 'makepath' to the create
call.  That has an impact on how the sharded nodes are created; it
seems we can't create "foo/0000000000", instead, we must always have
a prefix for the sequence, so we will end up with nodes like
"foo/seq00000000000".  That works well enough for this case.

Change-Id: I4e0e8ec579b291f4d410bcd95ac01f195b3007c1