Add item UUID to MQTT reporter
Since the MQTT reporter can be used to emit start or enqueue events,
it may be useful to match enqueue and dequeue events.  That could
be done with tenant+pipeline+change+patchset, but we also have a
UUID for queue items, so to make it simpler for MQTT consumers,
let's expose that.

Change-Id: Iff88bcfd73e00f292e0cc947f548582a276a7975