Use create_all for empty databases
Rather than running through all of the migrations when starting Zuul
with an empty database, this uses sqlalchemy's create_all method to
create it from the declarative schema.

To make sure that stays in sync with alembic, a test is added to run
DB creation both ways and compare.

The declaritive schema had one column with an incorrect type, and
several columns out of order; this adjusts the schema to match the
migrations.

Contrary to expectations, using sqlalchemy to create the schema actually
adds about 0.05 seconds on averate to test runtime.

Change-Id: I594b6980f5efa5fa4b8ca387c5d0ab4373b86394