Reset sys.last_traceback in run_handler
The variable sys.last_traceback holds the traceback of the last
occurred exception. This can block a large amount of memory for an
undefined amount of time (until the next exception is thrown). During
our attempts to hunt down a memory leak we always get this as the root
of a leaked pipeline or layout. While technically this is not a real
memory leak it still can block a large amount of memory and make
finding the really leaked objects harder.

Change-Id: Ic704affb0875bc972131510e82b57fb2f354c4dd