Switch from global to tenant event queues
With the way the global management and trigger event queues are
currently implemented we can't continue processing other events during a
reconfiguration and end up serializing reconfigurations.

The implementation is diverging from the spec here, as we did not
anticipate the need for tenant event queues back then.

We also add the following gauge metrics:
* zuul.tenant.<tenant>.trigger_events
* zuul.tenant.<tenant>.management_events

The `zuul.scheduler.eventqueues.trigger` gauge was removed together with
the global trigger event queue.

Change-Id: I1e0d872daf0fc4dd90e22b70c334e3e8152aa5b2